Split state into files but keep in same module
